🗃️ Committing everything that changed 🗃️
rootfs/usr/local/bin/entrypoint.sh
rootfs/usr/local/etc/docker/init.d/00-mysqld.sh
rootfs/usr/local/etc/docker/init.d/05-nginx.sh

+++ rootfs/usr/local/etc/docker/init.d/00-mysqld.sh
@@ -186,7 +186,7 @@ __run_pre_execute_checks() {
   # Put command to execute in parentheses
   {
     if [ ! -d "$DATABASE_DIR" ] || [ ! -f "$DATABASE_DIR/ibdata1" ]; then
-      rm -Rf "$DATABASE_DIR"
+      rm -Rf "${DATABASE_DIR:?}"/*
       mkdir -p "$DATABASE_DIR"
       chown -Rf $SERVICE_USER:$SERVICE_GROUP "$DATABASE_DIR"
       mysql_install_db --datadir=$DATABASE_DIR --user=$SERVICE_USER 2>/dev/null
